raise minimum and default alert delay for now
Android P power restrictions are causing problems with this due to a
much stricter implementation of App Standby. App Standby and Doze were
problems before but it got substantially worse. It could be worked
around using alarms and a battery optimization exception but for now the
alert delay needs to be significantly relaxed.