handle Pixel 3 TEE AVB key fingerprint change
The May 2019 update for the Pixel 3 and Pixel 3 XL moved to directly
using the sha256 hash of the verified boot public key as the verified
boot key fingerprint. It now matches the verified boot key fingerprint
provided by StrongBox.

Unfortunately, this is a hassle for AttestationServer since it pins the
verified boot key and needs to migrate to the new format. A comparable
change was not made for the Pixel 2 and Pixel 2 XL. It's likely that it
was done for consistency with the StrongBox implementation, which is
currently exclusive to the Pixel 3 and Pixel 3 XL among the supported
devices. It will hopefully not be changed for any other existing
devices. It would have been much harder to resolve this without access
to the devices and the ability to unlock them to flash factory images.