keep track of most recent login time
This also migrates older databases to using case insensitive usernames
like newer deployments. If there are duplicates, the migration will be
rolled back and the server won't start. Any duplicates will need to be
manually pruned before being able to migrate forward. It is impossible
to safely automate this and we don't want to indefinitely support case
sensitive usernames.