add backend support for longer certificate chains
This enables support for arbitrary length certificate chains by encoding
the certificate chain with the portable, versioned encoding used by our
attestation protocol. This does not yet add support to the frontend.

It compresses the certificate chain as a whole with a versioned static
dictionary so it will substantially reduce the size of the devices table
in the database. This won't currently result in a significant change in
the size of the overall database due to the inefficient text-based
storage for attestation history, but it will be helpful in the future
when AttestationServer is used at larger scales since attestation
history will be redone in a much more efficient way.

This will enable support for the attest key feature introduced in
Android 12 which adds an extra certificate to the chain. Hardware
support hasn't shipped yet but it likely will ship soon.

It would be possible to accomodate an extra certificate by adding a
nullable pinnedCertificate4 field. There has never been any actual
guarantee for the certificate chain length and there are other planned
changes which will increase the length in some cases. It's best to deal
with supporting arbitrary length chains now rather than putting it off
any longer.

The alternative would be making a normalized table:

    CREATE TABLE IF NOT EXISTS PinnedCertificates (
        fingerprint BLOB NOT NULL REFERENCES Devices (fingerprint) ON DELETE CASCADE,
        index INTEGER NOT NULL,
        certificate BLOB NOT NULL,
        PRIMARY KEY (fingerprint, index)
    )

There's no foreseeable reason to ever need queries of certificates
across multiple devices, so it doesn't make sense to have a table
dedicated to this. It would be a lot more complex and the compression
would either need to be done per-certificate or dropped.