more sensible exception for truncated cert chain
It makes more sense for it to fail signature verification rather than
an error from not having a valid root certificate. This made it harder
to figure out why the attest key feature wasn't working because the API
requires that you manually concatenate the certificate chains. It seemed
like it wasn't supported by the hardware when it was working as intended
because the exception we were throwing was misleading.