implement attest key downgrade support
Pixel 6 and Pixel 6 Pro have a bug making attest keys unusable after
upgrading. Detect this by catching IOException and downgrading to not
using an attest key for the case where we're using an existing attest
key instead of a newly generated one. For verification, we implement
support for allowing downgrading to verification based on the pinned
batch key chain via offsetting the pinning check. This preserves the
same security we had before attest keys while taking advantage of them
while they're working. The security level will automatically be reported
as Very High when the attest key is being used and will be downgraded to
High if the attest key isn't being used.

We can remove this after the issue is fixed.