enable SQLite mmap support
This significantly reduces memory usage by directly using the OS page
cache instead of using the per-connection SQLite page cache. It greatly
increases the displayed resident memory usage for the process since it
ends up counting the OS page cache but it can be easily verified that
the actual memory usage is lower.

It would be nice to set the mmap size value much higher, but contrary to
what's claimed by the documentation, setting this higher than the
configured maximum value will result in mmap not being used rather than
it being clamped to the max value. The reason for this is not obvious
from the SQLite implementation which matches what's documented but it
isn't what happens in practice via PRAGMA.