split out read transactions for email alerts
In order to avoid SQLITE_BUSY errors, we need to avoid ever upgrading
read transactions to write transactions. The usual way to do this is
wrapping it into an IMMEDIATE transaction to start out with a write
transaction. However, we don't want to make external connections during
a write transaction so the read transactions can be split out instead.