optimize alert dispatcher
This improves the account SELECT statement to only match accounts with
at least one expired device requiring an alert to be sent to avoid the
need to iterate over every single account checking for expired devices.

This also avoids loading every single account into an array and allows
us to trivially easily add a LIMIT clause to handle alerts in smaller
batches if that's ever needed.